stepik-django-main/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = ""
authors = [
    {name = "GanD0nI4e",email = "dimass004@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-filter (>=24.3,<25.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
